kevm-pyk/{__main__, utils}: pull out generic rpc_prove method
